Times can no longer be unsigned

+++ sbnobj/SBND/CRT/CRTCluster.hh
@@ -16,8 +16,8 @@ namespace sbnd::crt {
 
   class CRTCluster {
     
-    uint32_t  fTs0;          // T0 counter [ns]
-    uint32_t  fTs1;          // T1 counter [ns]
+    int64_t  fTs0;           // T0 counter [ns]
+    int64_t  fTs1;           // T1 counter [ns]
     uint32_t  fUnixS;        // Unixtime of event [s]
     uint16_t  fNHits;        // The number of strip hits forming the cluster
     CRTTagger fTagger;       // The tagger this cluster exists on
@@ -27,13 +27,13 @@ namespace sbnd::crt {
 
     CRTCluster();
     
-    CRTCluster(uint32_t _ts0, uint32_t _ts1, uint32_t _unixS, uint16_t _nHits, CRTTagger _tagger,
+    CRTCluster(int64_t _ts0, int64_t _ts1, uint32_t _unixS, uint16_t _nHits, CRTTagger _tagger,
                CoordSet _composition);
 
     virtual ~CRTCluster();
 
-    uint32_t  Ts0() const;
-    uint32_t  Ts1() const;
+    int64_t   Ts0() const;
+    int64_t   Ts1() const;
     uint32_t  UnixS() const;
     uint16_t  NHits() const;
     CRTTagger Tagger() const;
